restructured folder structure to allow for physics
