+++ nudge/include/Nudge/Maths/Quaternion.hpp
