+++ nudge/src/Maths/Matrix3.cpp
@@ -7,14 +7,13 @@
  * and linear algebra operations. The matrix uses column-major storage order.
  */
 
-#include "nudge/Matrix3.hpp"
+#include "nudge/Maths/Matrix3.hpp"
+#include "nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Matrix2.hpp"
+#include "Nudge/Maths/Vector2.hpp"
+#include "nudge/Maths/Vector3.hpp"
 
 #include <format>
-
-#include "nudge/MathF.hpp"
-#include "Nudge/Matrix2.hpp"
-#include "Nudge/Vector2.hpp"
-#include "nudge/Vector3.hpp"
 
 using std::runtime_error;
