+++ nudge/src/Maths/Matrix4.cpp
@@ -7,15 +7,13 @@
  * and linear algebra operations. The matrix uses column-major storage order.
  */
 
-#include "Nudge/Matrix4.hpp"
+#include "Nudge/Maths/Matrix4.hpp"
+#include "Nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Matrix3.hpp"
+#include "Nudge/Maths/Vector3.hpp"
+#include "Nudge/Maths/Vector4.hpp"
 
 #include <format>
-
-#include "Nudge/MathF.hpp"
-#include "Nudge/Matrix3.hpp"
-#include "Nudge/Vector2.hpp"
-#include "Nudge/Vector3.hpp"
-#include "Nudge/Vector4.hpp"
 
 using std::runtime_error;
