+++ nudge/src/Maths/Vector2.cpp
@@ -6,14 +6,13 @@
  * mathematical operations, utility functions, constructors, and operators.
  */
 
-#include "Nudge/Vector2.hpp"
+#include "Nudge/Maths/Vector2.hpp"
+#include "Nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Vector3.hpp"
+#include "Nudge/Maths/Vector4.hpp"
 
 #include <format>
 #include <stdexcept>
-
-#include "Nudge/MathF.hpp"
-#include "Nudge/Vector3.hpp"
-#include "Nudge/Vector4.hpp"
 
 using std::runtime_error;
