+++ tests/Maths/Matrix2Tests.cpp
@@ -3,9 +3,9 @@
 
 #include <gtest/gtest.h>
 
-#include "Nudge/MathF.hpp"
-#include "Nudge/Matrix2.hpp"
-#include "Nudge/Vector2.hpp"
+#include "Nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Matrix2.hpp"
+#include "Nudge/Maths/Vector2.hpp"
 
 using std::runtime_error;
 using std::numbers::pi_v;
