+++ tests/Maths/Vector2Tests.cpp
@@ -2,8 +2,8 @@
 
 #include <gtest/gtest.h>
 
-#include "Nudge/MathF.hpp"
-#include "Nudge/Vector2.hpp"
+#include "Nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Vector2.hpp"
 
 using std::runtime_error;
 using std::numbers::pi_v;
