+++ tests/Maths/Vector3Tests.cpp
@@ -2,9 +2,9 @@
 
 #include <gtest/gtest.h>
 
-#include "Nudge/MathF.hpp"
-#include "Nudge/Vector2.hpp"
-#include "Nudge/Vector3.hpp"
+#include "Nudge/Maths/MathF.hpp"
+#include "Nudge/Maths/Vector2.hpp"
+#include "Nudge/Maths/Vector3.hpp"
 
 using std::runtime_error;
 using std::numbers::pi_v;
